chore(artifacts): move Azure reference tests to the unit suite
The Azure reference tests build a local Artifact and mock the Azure
storage handler; they never touch the wandb backend, so they don't need
to run in the (slower) system-test job. Move them, together with the
mock_azure_handler and my_artifact fixtures they rely on, into a new
tests/unit_tests/test_artifacts/test_azure_references.py.

This is a pure relocation with no change in test behavior. The now-unused
`import unittest.mock` is dropped from the system test module.
